CLI: ArgumentParserの設定をサブコマンド用モジュールに分散配置するように修正
