+++ slack_message_analysis/cli.py
@@ -1,39 +1,24 @@
 from argparse import ArgumentParser
-from datetime import datetime
+import os
+from importlib import import_module
 
 
 def main() -> None:
-    from .collect import main as collect_main
-
     parser = ArgumentParser()
     subparsers = parser.add_subparsers()
+    excludes = set(['common.py', 'cli.py', 'models.py'])
+    topdir = os.path.dirname(__file__)
+    ns_root = os.path.dirname(topdir)
 
-    def setup_common_args(p: ArgumentParser) -> ArgumentParser:
-        # 共通の引数を設定します.
-        # 親parser.add_argumentではサブパーサのhelpに表示されないため
-        # サブパーサ毎に設定します
-        p.add_argument(
-            '--db', default='slack.sqlite',
-            help='SQLiteのパスを指定します。デフォルトはカレントディレクトリの"slack.sqlite"です')
-        return p
-
-    def datetime_parser(s: str) -> datetime:
-        return datetime.fromisoformat(s)
-
-    collect_parser = setup_common_args(subparsers.add_parser(
-        'collect', help='メッセージを収集しデータベースに格納します'))
-    collect_parser.add_argument(
-        '--token',
-        help='APIトークンを指定します。省略した場合はTOKEN環境変数の値が利用されます。')
-    collect_parser.add_argument(
-        '--since', help='メッセージ取得開始日時(ISO8601)を指定します。'
-        '省略した場合はDBに保存されている最新のメッセージ以降を取得対象とします。',
-        type=datetime_parser)
-    collect_parser.add_argument(
-        '--until', help='メッセージ取得終了日時(ISO8601)を指定します。'
-        '省略した場合はコマンド実行日の週の月曜日午前0時になります。',
-        type=datetime_parser)
-    collect_parser.set_defaults(func=collect_main)
+    for dirpath, _, filenames in os.walk(topdir):
+        ns = os.path.relpath(
+            dirpath, start=ns_root).replace('/', '.').replace('\\', '.') + '.'
+        for fn in filenames:
+            if not fn.endswith('.py') or fn.startswith('_') or fn in excludes:
+                continue
+            n, _ = os.path.splitext(fn)
+            m = import_module(ns + n)
+            getattr(m, 'init_argparser')(subparsers.add_parser)
 
     args = parser.parse_args()
     if hasattr(args, 'func'):
